refactor: extract create user form logic
